improved ui bugs. fixed ui based on remote config values. fixed popular movies list data in pagination.

+++ app/src/main/java/com/example/tmdb/model/DetailsResponses.kt
@@ -6,7 +6,7 @@ data class DetailsResponse(
 
     @SerializedName("backdrop_path") var backdropPath: String? = null,
     @SerializedName("genres") var genres: ArrayList<Genres> = arrayListOf(),
-    @SerializedName("id") var id: Int? = null,
+    @SerializedName("id") var id: Long? = null,
     @SerializedName("original_language") var originalLanguage: String? = null,
     @SerializedName("original_title") var originalTitle: String? = null,
     @SerializedName("overview") var overview: String? = null,
