+++ app/src/main/java/com/example/tmdb/model/PopularMoviesModel.kt
@@ -3,7 +3,7 @@ package com.example.tmdb.model
 import com.google.gson.annotations.SerializedName
 
 data class PopularMoviesModel(
-    @SerializedName("id") var popularMovieId: Int? = null,
+    @SerializedName("id") var popularMovieId: Long? = null,
     @SerializedName("poster_path") var posterPath: String? = null,
     @SerializedName("title") var popularMovieTitle: String? = null
 )
